requirements.txt
aim==3.29.1
albumentations==2.0.8
bitsandbytes==0.46.0
boto3==1.38.37
comet_ml==3.49.11
coremltools==8.3.0
cupy==8.3.0
datasets==3.1.0
dill==0.3.8
dvclive==3.48.2
evaluate==0.4.3
huggingface_hub==0.27.1
hydra-core==1.3.2
intel_extension_for_pytorch==2.7.0
ipython==8.12.3
kornia==0.7.3
lap==0.5.12
lib==4.0.0
lomo_optim==0.1.1
lvis==0.5.3
matplotlib==3.7.5
megatron==0.5.1
mmengine==0.10.5
mss==10.0.0
numpy==1.24.3
nvidia_ml_py==12.570.86
omegaconf==2.3.0
onnxruntime==1.19.2
onnxsim==0.4.36
opencv_python==4.10.0.84
openvino==2025.1.0
packaging==25.0
paddle==1.0.2
pafy==0.5.5
pandas==2.0.3
peft==0.15.2
Pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
pycocotools==2.0.7
pynvml==11.5.3
pytest==8.4.0
PyYAML==6.0.2
ray==2.10.0
Requests==2.32.4
safetensors==0.5.3
scikit_learn==1.3.2
scipy==1.15.3
seaborn==0.13.2
sentry_sdk==2.30.0
setuptools==60.2.0
skimage==0.0
super_gradients==3.7.1
tensorboardX==2.6.4
tensorflow==2.19.0
tensorflowjs==4.22.0
tensorrt==10.12.0.36
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
timm==1.0.15
torch==2.4.0
torch_npu==2.5.1
torchdata==0.11.0
torchpack==0.3.1
torchvision==0.19.0
transformers==4.46.3
triton==3.0.0
wandb==0.20.1
x2paddle==1.6.0
yt_dlp==2025.6.9